requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bootstrap4==0.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.2
django-bootstrap4==21.2
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-filter==21.1
django-import-export==2.7.1
django-tables2==2.4.1
environ==1.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
PyYAML==6.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.2.0
xlrd==2.0.1
xlwt==1.3.0
